requirements.txt
#func host start --useHttps --cert "server.pfx" --password "1111" --port 7443


aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
frozenlist==1.4.0
idna==3.6
multidict==6.0.4
openai==0.27.10
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
urllib3==2.1.0
yarl==1.9.3
azure-functions
